fix for mac unittests gif failure (#19)
* gif build error debug

* gif build error debug

* Update tests.json

* set external gif preference to off by default

+++ ci/before_install.sh
@@ -152,12 +152,12 @@ ln -sf artifacts/$TRAVIS_OS_NAME extlibs
 cd extlibs
 mkdir lib_orig
 cp -R lib/* lib_orig/.
-cd lib_orig
-ln -s libgif.7.dylib libgif.dylib
-cd ..
+if [ "$TRAVIS_OS_NAME" = "osx" ]; then
+rm -rf lib_orig/libgif.7.dylib
 rm -rf lib_orig/libpng.dylib
 rm -rf lib_orig/libsqlite3.dylib
 rm -rf lib_orig/libjpeg.dylib
+fi
 cd $TRAVIS_BUILD_DIR
 cd ..
 ln -sf rlExternals/Spark-Externals Spark-Externals
